scripts/ads/requirements.txt
# Minimal libs for Google Ads campaign creation
# Create a venv, then: pip install -r scripts/ads/requirements.txt

google-ads>=23.0.0
google-api-python-client>=2.140.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
requests>=2.32.0

